No conjugation processor; Some refactoring; Memory leak fixes

+++ src/dirichlet/solver/src/StreamParser.h
@@ -19,7 +19,7 @@ namespace Dirichlet {
 
         virtual Input parse_input() override;
 
-        double *parse_mesh(int width, int height);
+        double *parse_mesh(size_t width, size_t height);
 
     protected:
         istream &stream;
